+++ refena_flutter/lib/src/scope.dart
@@ -228,6 +228,13 @@ class RefenaScope extends StatefulWidget implements RefenaContainer {
   }
 
   @override
+  void rebuild<N extends RebuildableNotifier<T>, T>(
+    BaseProvider<N, T> provider,
+  ) {
+    _container.rebuild(provider);
+  }
+
+  @override
   void dispose<N extends BaseNotifier<T>, T>(BaseProvider<N, T> provider) {
     _container.dispose(provider);
   }